+++ sdk/core/core-lro/src/http/poller.ts
@@ -23,12 +23,37 @@ import { buildCreatePoller } from "../poller/poller.js";
  * Create a SimplePoller that will poll the LRO until it is done
  * Awaitting this would return an initialized poller
  */
-export async function createHttpSimplePoller<TResult, TState extends OperationState<TResult>>(
+export function createInitializedHttpPoller<TResult, TState extends OperationState<TResult>>(
   lro: LongRunningOperation,
   options?: CreateHttpPollerOptions<TResult, TState>,
-): Promise<SimplePollerLike<TState, TResult>> {
-  const poller = createHttpPoller(lro, "SimplePoller", options);
-  // return the poller after it has been initialized
+): Promise<PollerLike<TState, TResult>>;
+/**
+ * Explicitly specify the type of poller as "Poller" to create
+ */
+export function createInitializedHttpPoller<TResult, TState extends OperationState<TResult>>(
+  lro: LongRunningOperation,
+  type: "Poller",
+  options?: CreateHttpPollerOptions<TResult, TState>,
+): Promise<PollerLike<TState, TResult>>;
+/**
+ * Explicitly specify the type of poller as "SimplePoller" to create
+ */
+export function createInitializedHttpPoller<TResult, TState extends OperationState<TResult>>(
+  lro: LongRunningOperation,
+  type: "SimplePoller",
+  options?: CreateHttpPollerOptions<TResult, TState>,
+): Promise<SimplePollerLike<TState, TResult>>;
+export async function createInitializedHttpPoller<TResult, TState extends OperationState<TResult>>(
+  lro: LongRunningOperation,
+  typeOrOptions: "Poller" | "SimplePoller" | CreateHttpPollerOptions<TResult, TState> = "Poller",
+  options?: CreateHttpPollerOptions<TResult, TState>,
+): Promise<SimplePollerLike<TState, TResult> | PollerLike<TState, TResult>> {
+  const type = typeof typeOrOptions === "string" ? typeOrOptions : "Poller";
+  options = typeof typeOrOptions === "string" ? options : typeOrOptions;
+  const poller =
+    type === "Poller"
+      ? createHttpPoller(lro, options)
+      : createHttpPoller(lro, "SimplePoller", options);
   await poller.submitted();
   return poller;
 }